% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{train}
\alias{train}
\title{train entrena la red neuronal usando una topologia de arbol}
\usage{
train(numberOfChildrenperNode, treeHeight, initialLearningRate,
  finalLearningRate, initialRadius, finalRadius, numberOfIterations, data)
}
\arguments{
\item{numberOfChildrenperNode}{A integer number}

\item{treeHeight}{A integer number}

\item{initialLearningRate}{A float number}

\item{finalLearningRate}{A float number}

\item{initialRadius}{A integer number}

\item{finalRadius}{A integer number}

\item{numberOfIterations}{A integer number}

\item{data}{A data frame}
}
\value{
La red neuronal entrenada con topologia de un arbol k ario completo
}
\description{
train entrena la red neuronal usando una topologia de arbol
}
\examples{
library(ggplot2)
##################### EXAMPLE 1 : IRIS DATASET
###parameters
numberOfIterations <- 600000
initialLearningRate <- 1
finalLearningRate<- 0
initialRadius <- 7
finalRadius <- 1
numberOfChildrenperNode <- 2
treeHeight <- 3

##training phase
data(iris)
data<-iris[-5] ## load a dataset
ti <- proc.time() # start watch
neurons <- train(numberOfChildrenperNode,treeHeight,initialLearningRate,finalLearningRate,initialRadius,finalRadius,numberOfIterations, data)
tf <-proc.time()    # stop watch
tf-ti #print execution time

##visualization phase
graficar(data,neurons,numberOfChildrenperNode) #plot the scatter plot
}
